+++ src/backend/core/src/database/red_cap/participants/extra.rs
@@ -6,21 +6,21 @@ use sqlx::prelude::FromRow;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, FromRow)]
 pub struct ParticipantEmergencyContact {
-    pub id: i64,
-    pub participant_id: i64,
+    pub id: i32,
+    pub participant_id: i32,
 }
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, FromRow)]
 pub struct ParticipantHealthCareProviders {
-    pub id: i64,
-    pub participant_id: i64,
+    pub id: i32,
+    pub participant_id: i32,
 }
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, FromRow)]
 pub struct ParticipantPharmacy {
-    pub id: i64,
-    pub participant_id: i64,
+    pub id: i32,
+    pub participant_id: i32,
 }
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, FromRow)]
 pub struct ParticipantAdvancedCarePlan {
-    pub id: i64,
-    pub participant_id: i64,
+    pub id: i32,
+    pub participant_id: i32,
 }
